requirements.txt
agora_python_server_sdk==2.2.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
numpy==2.3.1
paho-mqtt==2.1.0
propcache==0.3.2
python-dotenv==1.1.0
websockets==15.0.1
yarl==1.20.1
